farmedFish 진행중, config파일에 camelcase적용

+++ src/main/java/kr/co/ffm/system/farmedfish/FarmedFish.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class FarmedFish {
     private int no;
     private String name;
-    private int temperature;
-    private int ph;
-    private int oxygen;
+    private double temperature;
+    private double ph;
+    private double oxygen;
     private String feedingTime;
     private int feedingAmount;
     private String active;
